--- src/components/pagination/style.css
@@ -1,37 +1,0 @@
-.pagination-container {
-    color: var(--theme-dark-gray-text);
-    padding: 10px;
-    align-self: flex-end;
-    max-width: 150px;
-}
-
-.pagination-btn {
-    user-select: none;
-    color: #8797AD;
-    width: 20px;
-}
-
-.pagination-btn.disabled {
-    pointer-events: none;
-}
-
-.pagination-btn > div {
-    width: 100%;
-}
-
-.arrow.active:hover {
-    cursor: pointer;
-    background-color: rgba(0,0,0,0.1);
-}
-
-.arrow {
-    color: #8797AD
-}
-.arrow.active {
-    color: #3E94C5
-}
-
-.active-pagination-btn {
-    background-color: var(--theme-brand-blue);
-    color: white;
-}
