+++ src/layouts/events/index.js
@@ -1,21 +1,21 @@
-import React, { useCallback, useState } from 'react';
-import { Config } from '../../util';
-import {  VscCloud, VscPlay, VscDebugStepInto } from 'react-icons/vsc';
+import { useEvents } from 'direktiv-react-hooks';
+import { useState } from 'react';
+import { VscCloud, VscDebugStepInto, VscPlay } from 'react-icons/vsc';
+import { AutoSizer } from 'react-virtualized';
 import Button from '../../components/button';
 import ContentPanel, { ContentPanelBody, ContentPanelHeaderButton, ContentPanelTitle, ContentPanelTitleIcon } from '../../components/content-panel';
-import FlexBox from '../../components/flexbox';
-import {useEvents} from 'direktiv-react-hooks'
-import Modal, { ButtonDefinition } from '../../components/modal';
 import DirektivEditor from '../../components/editor';
-import { AutoSizer } from 'react-virtualized';
+import FlexBox from '../../components/flexbox';
 import HelpIcon from "../../components/help";
+import Modal, { ButtonDefinition } from '../../components/modal';
+import { Config } from '../../util';
 
-import * as dayjs from "dayjs"
+import * as dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
-import utc from "dayjs/plugin/utc"
-import './style.css'
+import utc from "dayjs/plugin/utc";
 import { Link } from 'react-router-dom';
-import Pagination from '../../components/pagination';
+import Pagination, { usePageHandler } from '../../components/pagination';
+import './style.css';
 
 dayjs.extend(utc)
 dayjs.extend(relativeTime);
@@ -43,152 +43,143 @@ function EventsPage(props) {
     let {namespace} = props;
 
     // errHistory and errListeners TODO show error if one
-    const [listenersParam, setListenersParam] = useState([`limit=${PAGE_SIZE}`])
-    const [historyParam, setHistoryParam] = useState([`limit=${PAGE_SIZE}`])
 
-    let {eventHistory, eventListeners, eventListenersTotalCount, eventListenersPageInfo, eventHistoryTotalCount, eventHistoryPageInfo, sendEvent, replayEvent} = useEvents(Config.url, true, namespace, localStorage.getItem("apikey"), {listeners: listenersParam, history: historyParam})
-    const updateEventHistoryPage = useCallback((newParam)=>{
-        setHistoryParam([...newParam])
-    }, [])
-    const updateListenersPage = useCallback((newParam)=>{
-        setListenersParam([...newParam])
-    }, [])
-    return(
+    const historyPageHandler = usePageHandler(PAGE_SIZE)
+    const listenersPageHandler = usePageHandler(PAGE_SIZE)
+
+    let {eventHistory, eventListeners, eventListenersPageInfo, eventHistoryPageInfo, sendEvent, replayEvent} = useEvents(Config.url, true, namespace, localStorage.getItem("apikey"), {listeners: [listenersPageHandler.pageParams], history: [historyPageHandler.pageParams]})
+    return (
         <>
-            <FlexBox className="gap col" style={{paddingRight: "8px"}}>
+            <FlexBox className="gap col" style={{ paddingRight: "8px" }}>
                 <FlexBox>
                     <ContentPanel style={{ width: "100%" }}>
                         <ContentPanelTitle>
                             <ContentPanelTitleIcon>
-                                <VscCloud/>
+                                <VscCloud />
                             </ContentPanelTitleIcon>
-                            <FlexBox style={{display:"flex", alignItems:"center"}} className="gap">
+                            <FlexBox style={{ display: "flex", alignItems: "center" }} className="gap">
                                 <div>
                                     Cloud Events History
                                 </div>
                                 <HelpIcon msg={"A history of events that have hit this specific namespace."} />
                             </FlexBox>
-                            <SendEventModal sendEvent={sendEvent}/>
+                            <SendEventModal sendEvent={sendEvent} />
                         </ContentPanelTitle>
                         <ContentPanelBody>
-                            <div style={{maxHeight: "40vh", overflowY: "auto", fontSize: "12px", minWidth: "100%"}}>
-                                <table className="cloudevents-table" style={{minWidth: "440px", width: "100%"}}>
-                                    <thead>
-                                        <tr>
-                                            <th>
-                                                Type
-                                            </th>
-                                            <th style={{width:"250px"}}>
-                                                Source
-                                            </th>
-                                            <th>
-                                                Time
-                                            </th>
-                                            <th style={{textAlign:'center'}}>
-                                                  Actions
-                                            </th>
-                                        </tr>
-                                    </thead>
-                                    {eventHistory !== null && typeof eventHistory === typeof [] && eventHistory.length > 0 ?
-                                    <tbody>
-                                        {eventHistory.map((obj) => {
-                                            return <tr style={{ borderBottom: "1px solid #f4f4f4" }}>
-                                                <td title={obj.type} style={{ textOverflow: "ellipsis", overflow: "hidden" }}>
-                                                    {obj.type}
-                                                </td>
-                                                <td title={obj.source} style={{ textOverflow: "ellipsis", overflow: "hidden" }}>
-                                                    {obj.source}
-                                                </td>
-                                                <td>
-                                                    {dayjs.utc(obj.receivedAt).local().format("HH:mm:ss a")}
-                                                </td>
-                                                <td style={{ textAlign: 'center', justifyContent: "center", }}>
-                                                    <FlexBox className={"gap center"}>
-                                                        <Modal
-                                                            className="run-workflow-modal"
-                                                            style={{ justifyContent: "flex-end" }}
-                                                            modalStyle={{ color: "black", width: "360px" }}
-                                                            title="Retrigger Event"
-                                                            onClose={() => {
-                                                            }}
-                                                            btnStyle={{ width: "unset" }}
-                                                            button={
-                                                                <Button className="small light bold" tip="Retrigger Event">
-                                                                    <VscPlay /> <span className='hide-800'>Retrigger</span>
-                                                                </Button>
-                                                            }
-                                                            actionButtons={[
-                                                                ButtonDefinition("Retrigger", async () => {
-                                                                    await replayEvent(obj.id)
-                                                                }, "small", () => { }, true, true),
-                                                                ButtonDefinition("Cancel", async () => {
-                                                                }, "small light", () => { }, true, false)
-                                                            ]}
-                                                        >
-                                                            <FlexBox style={{ overflow: "hidden" }}>
-                                                                Are you sure you want to retrigger {obj.id}?
-                                                            </FlexBox>
-                                                        </Modal>
-                                                        <Modal
-                                                            className="run-workflow-modal"
-                                                            modalStyle={{ color: "black", minWidth: "360px", width: "50vw", height: "40vh", minHeight: "680px" }}
-                                                            title="View Event"
-                                                            onClose={() => {
-                                                            }}
-                                                            btnStyle={{ width: "unset" }}
-                                                            button={
-                                                                <Button className="small light bold">
-                                                                    View
-                                                                </Button>}
-                                                            actionButtons={[
-                                                                ButtonDefinition("Close", async () => {
-                                                                }, "small light", () => { }, true, false)
-                                                            ]}
-                                                        >
-                                                            <FlexBox className="col" style={{ overflow: "hidden" }}>
-                                                                <AutoSizer>
-                                                                    {({ height, width }) => (
-                                                                        <DirektivEditor noBorderRadius value={atob(obj.cloudevent)} readonly={true} dlang="plaintext"
-                                                                            options={{
-                                                                                autoLayout: true
-                                                                            }}
-                                                                            width={width}
-                                                                            height={height}
-                                                                        />
-                                                                    )}
-                                                                </AutoSizer>
-                                                            </FlexBox>
-                                                        </Modal>
-                                                    </FlexBox>
-                                                </td>
+                            <FlexBox className="col" style={{ justifyContent: "space-between" }}>
+                                <div style={{ maxHeight: "40vh", overflowY: "auto", fontSize: "12px", minWidth: "100%" }}>
+                                    <table className="cloudevents-table" style={{ minWidth: "440px", width: "100%" }}>
+                                        <thead>
+                                            <tr>
+                                                <th>
+                                                    Type
+                                                </th>
+                                                <th style={{ width: "250px" }}>
+                                                    Source
+                                                </th>
+                                                <th>
+                                                    Time
+                                                </th>
+                                                <th style={{ textAlign: 'center' }}>
+                                                    Actions
+                                                </th>
                                             </tr>
-                                        })}
-                                    </tbody> : 
-                                    <FlexBox className='table-no-content'>
-                                        No cloud events history
-                                    </FlexBox>
-                                }
-                                </table>
-                            </div>
-                            {
-                                !!eventHistoryTotalCount && 
-                                <Pagination
-                                    pageSize={PAGE_SIZE}
-                                    total={eventHistoryTotalCount}
-                                    pageInfo={eventHistoryPageInfo}
-                                    updatePage={updateEventHistoryPage}
-                                />
-                            }
-                        </ContentPanelBody>                 
+                                        </thead>
+                                        {eventHistory !== null && typeof eventHistory === typeof [] && eventHistory.length > 0 ?
+                                            <tbody>
+                                                {eventHistory.map((obj) => {
+                                                    return <tr style={{ borderBottom: "1px solid #f4f4f4" }}>
+                                                        <td title={obj.type} style={{ textOverflow: "ellipsis", overflow: "hidden" }}>
+                                                            {obj.type}
+                                                        </td>
+                                                        <td title={obj.source} style={{ textOverflow: "ellipsis", overflow: "hidden" }}>
+                                                            {obj.source}
+                                                        </td>
+                                                        <td>
+                                                            {dayjs.utc(obj.receivedAt).local().format("HH:mm:ss a")}
+                                                        </td>
+                                                        <td style={{ textAlign: 'center', justifyContent: "center", }}>
+                                                            <FlexBox className={"gap center"}>
+                                                                <Modal
+                                                                    className="run-workflow-modal"
+                                                                    style={{ justifyContent: "flex-end" }}
+                                                                    modalStyle={{ color: "black", width: "360px" }}
+                                                                    title="Retrigger Event"
+                                                                    onClose={() => {
+                                                                    }}
+                                                                    btnStyle={{ width: "unset" }}
+                                                                    button={
+                                                                        <Button className="small light bold" tip="Retrigger Event">
+                                                                            <VscPlay /> <span className='hide-800'>Retrigger</span>
+                                                                        </Button>
+                                                                    }
+                                                                    actionButtons={[
+                                                                        ButtonDefinition("Retrigger", async () => {
+                                                                            await replayEvent(obj.id)
+                                                                        }, "small", () => { }, true, true),
+                                                                        ButtonDefinition("Cancel", async () => {
+                                                                        }, "small light", () => { }, true, false)
+                                                                    ]}
+                                                                >
+                                                                    <FlexBox style={{ overflow: "hidden" }}>
+                                                                        Are you sure you want to retrigger {obj.id}?
+                                                                    </FlexBox>
+                                                                </Modal>
+                                                                <Modal
+                                                                    className="run-workflow-modal"
+                                                                    modalStyle={{ color: "black", minWidth: "360px", width: "50vw", height: "40vh", minHeight: "680px" }}
+                                                                    title="View Event"
+                                                                    onClose={() => {
+                                                                    }}
+                                                                    btnStyle={{ width: "unset" }}
+                                                                    button={
+                                                                        <Button className="small light bold">
+                                                                            View
+                                                                        </Button>}
+                                                                    actionButtons={[
+                                                                        ButtonDefinition("Close", async () => {
+                                                                        }, "small light", () => { }, true, false)
+                                                                    ]}
+                                                                >
+                                                                    <FlexBox className="col" style={{ overflow: "hidden" }}>
+                                                                        <AutoSizer>
+                                                                            {({ height, width }) => (
+                                                                                <DirektivEditor noBorderRadius value={atob(obj.cloudevent)} readonly={true} dlang="plaintext"
+                                                                                    options={{
+                                                                                        autoLayout: true
+                                                                                    }}
+                                                                                    width={width}
+                                                                                    height={height}
+                                                                                />
+                                                                            )}
+                                                                        </AutoSizer>
+                                                                    </FlexBox>
+                                                                </Modal>
+                                                            </FlexBox>
+                                                        </td>
+                                                    </tr>
+                                                })}
+                                            </tbody> :
+                                            <FlexBox className='table-no-content'>
+                                                No cloud events history
+                                            </FlexBox>
+                                        }
+                                    </table>
+                                </div>
+                                <FlexBox className="row" style={{justifyContent:"flex-end", paddingBottom:"1em", flexGrow: 0}}>
+                                    <Pagination pageHandler={historyPageHandler} pageInfo={eventHistoryPageInfo} />
+                                </FlexBox>
+                            </FlexBox>
+                        </ContentPanelBody>
                     </ContentPanel>
                 </FlexBox>
                 <FlexBox>
                     <ContentPanel style={{ width: "100%" }}>
                         <ContentPanelTitle>
                             <ContentPanelTitleIcon>
-                                <VscDebugStepInto/>
+                                <VscDebugStepInto />
                             </ContentPanelTitleIcon>
-                            <FlexBox style={{display:"flex", alignItems:"center"}} className="gap">
+                            <FlexBox style={{ display: "flex", alignItems: "center" }} className="gap">
                                 <div>
                                     Active Event Listeners
                                 </div>
@@ -196,70 +187,66 @@ function EventsPage(props) {
                             </FlexBox>
                         </ContentPanelTitle>
                         <ContentPanelBody>
-                            <div style={{maxHeight: "40vh", overflowY: "auto", fontSize: "12px"}}>
-                                <table className="event-listeners-table" style={{width: "100%"}}>
-                                    <thead>
-                                        <tr>
-                                            <th>
-                                                Workflow
-                                            </th>
-                                            <th>
-                                                Type
-                                            </th>
-                                            <th>
-                                                Mode
-                                            </th>
-                                            <th>
-                                                Updated
-                                            </th>
-                                            <th>
-                                                Event Types
-                                            </th>
-                                        </tr>
-                                    </thead>
-                                    {eventListeners !== null && typeof eventListeners === typeof [] && eventListeners?.length > 0 ?
-                                    <tbody>
-                                        {eventListeners.map((obj)=>{
-                                            return(
-                                                <tr  style={{borderBottom:"1px solid #f4f4f4"}}>
-                                                    <td style={{textOverflow:"ellipsis", overflow:"hidden"}}>
-                                                        <Link style={{color:"#2396d8"}} to={`/n/${namespace}/explorer${obj.workflow}`}>
-                                                            {obj.workflow}
-                                                        </Link> 
-                                                    </td>
-                                                    <td style={{textOverflow:"ellipsis", overflow:"hidden"}}>
-                                                        {obj.instance !== "" ? <Link style={{color:"#2396d8"}} to={`/n/${namespace}/instances/${obj.instance}`}>{obj.instance.split("-")[0]}</Link> : "start"}
-                                                    </td>
-                                                    <td style={{textOverflow:"ellipsis", overflow:"hidden"}}>
-                                                        {obj.mode}
-                                                    </td>
-                                                    <td>
-                                                        {dayjs.utc(obj.updatedAt).local().format("HH:mm:ss a")}
-                                                    </td>
-                                                    <td className="event-split">
-                                                        {obj.events.map((obj)=>{
-                                                            return <span>{obj.type}</span>
-                                                        })}
-                                                    </td>
-                                                </tr>
-                                            )
-                                        })}
-                                    </tbody>:
-                                        <FlexBox className='table-no-content'>
-                                            No active event listeners
-                                        </FlexBox>
-                                    }
-                                </table>
-                            </div>
-                            {
-                                !!eventListenersTotalCount && 
-                                <Pagination
-                                    pageSize={PAGE_SIZE}
-                                    total={eventListenersTotalCount}
-                                    pageInfo={eventListenersPageInfo}
-                                    updatePage={updateListenersPage}
-                                />
-                            }
+                            <FlexBox className="col" style={{ justifyContent: "space-between" }}>
+                                <div style={{ maxHeight: "40vh", overflowY: "auto", fontSize: "12px" }}>
+                                    <table className="event-listeners-table" style={{ width: "100%" }}>
+                                        <thead>
+                                            <tr>
+                                                <th>
+                                                    Workflow
+                                                </th>
+                                                <th>
+                                                    Type
+                                                </th>
+                                                <th>
+                                                    Mode
+                                                </th>
+                                                <th>
+                                                    Updated
+                                                </th>
+                                                <th>
+                                                    Event Types
+                                                </th>
+                                            </tr>
+                                        </thead>
+                                        {eventListeners !== null && typeof eventListeners === typeof [] && eventListeners?.length > 0 ?
+                                            <tbody>
+                                                {eventListeners.map((obj) => {
+                                                    return (
+                                                        <tr style={{ borderBottom: "1px solid #f4f4f4" }}>
+                                                            <td style={{ textOverflow: "ellipsis", overflow: "hidden" }}>
+                                                                <Link style={{ color: "#2396d8" }} to={`/n/${namespace}/explorer${obj.workflow}`}>
+                                                                    {obj.workflow}
+                                                                </Link>
+                                                            </td>
+                                                            <td style={{ textOverflow: "ellipsis", overflow: "hidden" }}>
+                                                                {obj.instance !== "" ? <Link style={{ color: "#2396d8" }} to={`/n/${namespace}/instances/${obj.instance}`}>{obj.instance.split("-")[0]}</Link> : "start"}
+                                                            </td>
+                                                            <td style={{ textOverflow: "ellipsis", overflow: "hidden" }}>
+                                                                {obj.mode}
+                                                            </td>
+                                                            <td>
+                                                                {dayjs.utc(obj.updatedAt).local().format("HH:mm:ss a")}
+                                                            </td>
+                                                            <td className="event-split">
+                                                                {obj.events.map((obj) => {
+                                                                    return <span>{obj.type}</span>
+                                                                })}
+                                                            </td>
+                                                        </tr>
+                                                    )
+                                                })}
+                                            </tbody> :
+                                            <FlexBox className='table-no-content'>
+                                                No active event listeners
+                                            </FlexBox>
+                                        }
+                                    </table>
+                                </div>
+                                <FlexBox className="row" style={{justifyContent:"flex-end", paddingBottom:"1em", flexGrow: 0}}>
+                                    <Pagination pageHandler={listenersPageHandler} pageInfo={eventListenersPageInfo} />
+                                </FlexBox>
+                            </FlexBox>
                         </ContentPanelBody>
                     </ContentPanel>
                 </FlexBox>
